fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@17064 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/windows/GUISUMOAbstractView.cpp
@@ -358,7 +358,7 @@ GUISUMOAbstractView::getObjectsInBoundary(const Boundary& bound) {
     for (int i = 0; i < nb_hits; ++i) {
         numNames = *ptr;
         ptr += 3;
-        for (int j = 0; j < numNames; j++) {     /*  for each name */
+        for (int j = 0; j < (int)numNames; j++) {
             result.push_back(*ptr);
             ptr++;
         }
